# Optimize ByteBuf.setCharSequence for adaptive allocator