@chrisvest 2025-05-14
Benchmark numbers from my machine:

```
Before:
Benchmark                                                   (bufferType)  (charsetName)  (size)   Mode  Cnt          Score         Error  Units
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII       8  thrpt   20  131018022.048 ± 1023708.262  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII      64  thrpt   20   41468930.766 ± 1082323.255  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII    1024  thrpt   20    4691333.194 ±   14046.115  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII       8  thrpt   20   79807673.356 ±  524191.202  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII      64  thrpt   20   15016758.610 ±   38431.438  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII    1024  thrpt   20     990789.479 ±    2353.519  ops/s

After:
Benchmark                                                   (bufferType)  (charsetName)  (size)   Mode  Cnt          Score         Error  Units
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII       8  thrpt   20  126180019.645 ± 1119409.293  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII      64  thrpt   20   41367018.242 ±  437086.137  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence    ADAPTIVE_HEAD       US-ASCII    1024  thrpt   20    4665530.303 ±   12217.104  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII       8  thrpt   20   99074464.746 ±  583860.874  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII      64  thrpt   20   21527200.403 ±   63990.160  ops/s
AbstractByteBufGetCharSequenceBenchmark.setCharSequence  ADAPTIVE_DIRECT       US-ASCII    1024  thrpt   20    1487919.795 ±   11867.385  ops/s
```